Finish WIP impl. of CheckDays
The AbuseIPDB API is a bit annoying. If there's a single report for an
IP it returns an object representing that report. If there are multiple
reports, it returns an array of objects. Would have been much simpler to
implement if it had just always returned an array of objects, even when
there's only one object in the array.