Refactor error handling
The API returns 400 level HTTP status codes if the request is incorrect.
It returns a JSON response that could be parsed, but I'm kind of
hesitant when I'm not sure of how exactly the format is defined. For
now, let the user deal with it: put it in a RequestError and let them
parse it if they want.